backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-bcrypt
tensorflow-macos==2.9.0
tensorflow-metal==0.5.0
Pillow
